Update parameter handling, add Morse, general cleanup
Added the functionality to specify the species that a potential will compute over. This allows for the definition of different element-element interactions. This also places nicely with MixedPotential.

Added functions to get and handle potential parameters. There is still some additional features to add (potentially only deserializing a subset of parameters?)

Added the Morse potential and related basic methods

Removed GaN information because it is outdated and is replaced with species handling.

Refactored utilities, neighborlist

ACE is still being withheld because of issues with ACE1.jl compatability and CESMIX registry.

Co-Authored-By: Jeremiah DeGreeff <jrdegreeff@gmail.com>