Update ACE, SNAP,  add LinearBasisPotentials and Neural Network BasisPotentials
In preparation for some clean up, I have restructured the code base significant. ACE and SNAP now refer to BasisSystems, methods that compute local and force descriptors. The potentials are now classified as methods that take in a configuration and a basis system and return an energy, force, or virial. New examples of BasisPotentials include the LinearBasisPotential (which used to be referred to as ACE and SNAP) as well as the NeuralNetworkBasisPotential, that uses a neural network.